+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## Changelog
 
 ### Version 0.0.1
+
+- Created SvelteKit starter
+- Created Angular starter